+++ pkg/scoringitem/scoring_item.go
@@ -1,10 +1,10 @@
 package scoringitem
 
 type ScoringItem struct {
-	IsReverseItem   bool           `json:"isReverseItem"`
-	LeagueRanking   int            `json:"leagueRanking"`
-	LeagueTotal     int            `json:"leagueTotal"`
-	Points          int            `json:"points"`
-	PointsOverrides map[string]int `json:"pointsOverrides"`
-	StatId          int            `json:"statId"`
+	IsReverseItem   bool               `json:"isReverseItem"`
+	LeagueRanking   float32            `json:"leagueRanking"`
+	LeagueTotal     float32            `json:"leagueTotal"`
+	Points          float32            `json:"points"`
+	PointsOverrides map[string]float32 `json:"pointsOverrides"`
+	StatId          int                `json:"statId"`
 }
